fixes for maya 2022, using the new folder structure to maya app folder (the new structure follows what's inside the installation rpm)
using macfix() method, which essentially allows to prepend a path to the root app of an app. Originally written for OSX, works perfectly to allow maya to retain the original folder structure it has inside the RPM. This will make much easier to install maya from now on, just extracting the files from the RPM and putting then inside the maya/VERSION/ folder.

Also, we now make use of MAYA_PYTHON_VERSION, to set the version of python maya should use. For now, hardcoded to 2, since PipeVFX only supports python 2 for now.